Use uint32 instead of BigInt
We're working towards a SPDZ2k implementation with k=32

+++ src/NiMPC/Communication.nim
@@ -1,10 +1,9 @@
 import tables
-import bigints
 import asyncdispatch
 include Parties
 
 type
-  Messages = FutureStream[BigInt]
+  Messages = FutureStream[uint32]
   Inbox = ref Table[Party, Messages]
   Inboxes = ref Table[Party, Inbox]
 
@@ -13,15 +12,15 @@ var inboxes = Inboxes()
 proc inbox(party: Party): Inbox =
   inboxes.mgetOrPut(party, Inbox())
 
-method send*(sender: Party, recipient: Party, value: BigInt): Future[void] {.async,base.} =
-  let messages = recipient.inbox.mgetOrPut(sender, newFutureStream[BigInt]())
+method send*(sender: Party, recipient: Party, value: uint32): Future[void] {.async,base.} =
+  let messages = recipient.inbox.mgetOrPut(sender, newFutureStream[uint32]())
   await messages.write(value)
 
-method broadcast*(sender: Party, value: BigInt): Future[void] {.async,base.} =
+method broadcast*(sender: Party, value: uint32): Future[void] {.async,base.} =
   for recipient in sender.peers:
     await sender.send(recipient, value)
 
-method receive*(recipient: Party, sender: Party): Future[BigInt] {.async,base.} =
-  let messages = recipient.inbox.mgetOrPut(sender, newFutureStream[BigInt]())
+method receive*(recipient: Party, sender: Party): Future[uint32] {.async,base.} =
+  let messages = recipient.inbox.mgetOrPut(sender, newFutureStream[uint32]())
   let (_, received) = await messages.read()
   result = received
